[parallel] add base class for IPythonPool and MPIPool
