+++ src/pymor/parallel/mpi.py
@@ -5,26 +5,17 @@
 from __future__ import absolute_import, division, print_function
 
 from itertools import chain
-import weakref
 
 
-from pymor.core.interfaces import ImmutableInterface
-from pymor.core.pickle import FunctionPicklingWrapper
-from pymor.parallel.defaultimpl import WorkerPoolDefaultImplementations
-from pymor.parallel.interfaces import WorkerPoolInterface, RemoteObjectInterface
-from pymor.tools.counter import Counter
+from pymor.parallel.basic import WorkerPoolBase
 from pymor.tools import mpi
 
 
-FunctionType = type(lambda x: x)
-
-
-class MPIPool(WorkerPoolDefaultImplementations, WorkerPoolInterface):
+class MPIPool(WorkerPoolBase):
 
     def __init__(self):
+        super(MPIPool, self).__init__()
         self.logger.info('Connected to {} ranks'.format(mpi.size))
-        self._pushed_immutable_objects = {}
-        self._remote_objects_created = Counter()
         self._payload = mpi.call(mpi.function_call_manage, _setup_worker)
 
     def __del__(self):
@@ -33,35 +24,13 @@ class MPIPool(WorkerPoolDefaultImplementations, WorkerPoolInterface):
     def __len__(self):
         return mpi.size
 
-    def push(self, obj):
-        if isinstance(obj, ImmutableInterface):
-            uid = obj.uid
-            if uid not in self._pushed_immutable_objects:
-                remote_id = mpi.call(mpi.function_call_manage, _push_object, obj)
-                self._pushed_immutable_objects[uid] = (remote_id, 1)
-            else:
-                remote_id, ref_count = self._pushed_immutable_objects[uid]
-                self._pushed_immutable_objects[uid] = (remote_id, ref_count + 1)
-            return MPIRemoteObject(self, remote_id, uid=uid)
-        else:
-            remote_id = mpi.call(mpi.function_call_manage, _push_object, obj)
-            return MPIRemoteObject(self, remote_id)
+    def _push_object(self, obj):
+        return mpi.call(mpi.function_call_manage, _push_object, obj)
 
-    def _map_kwargs(self, kwargs):
-        pushed_immutable_objects = self._pushed_immutable_objects
-        return {k: (pushed_immutable_objects.get(v.uid, (v, 0))[0] if isinstance(v, ImmutableInterface) else
-                    v.remote_id if isinstance(v, MPIRemoteObject) else
-                    v)
-                for k, v in kwargs.iteritems()}
-
-    def apply(self, function, *args, **kwargs):
-        function = FunctionPicklingWrapper(function)
-        kwargs = self._map_kwargs(kwargs)
+    def _apply(self, function, *args, **kwargs):
         return mpi.call(mpi.function_call, _worker_call_function, function, *args, **kwargs)
 
-    def apply_only(self, function, worker, *args, **kwargs):
-        function = FunctionPicklingWrapper(function)
-        kwargs = self._map_kwargs(kwargs)
+    def _apply_only(self, function, worker, *args, **kwargs):
         payload = mpi.get_object(self._payload)
         payload[0] = (function, args, kwargs)
         try:
@@ -70,35 +39,17 @@ class MPIPool(WorkerPoolDefaultImplementations, WorkerPoolInterface):
             payload[0] = None
         return result
 
-    def map(self, function, *args, **kwargs):
-        function = FunctionPicklingWrapper(function)
-        kwargs = self._map_kwargs(kwargs)
+    def _map(self, function, chunks, **kwargs):
         payload = mpi.get_object(self._payload)
-        payload[0] = _split_into_chunks(mpi.size, *args)
+        payload[0] = chunks
         try:
             result = mpi.call(mpi.function_call, _worker_map_function, self._payload, function, **kwargs)
         finally:
             payload[0] = None
         return result
 
-
-class MPIRemoteObject(RemoteObjectInterface):
-
-    def __init__(self, pool, remote_id, uid=None):
-        self.pool = weakref.ref(pool)
-        self.remote_id = remote_id
-        self.uid = uid
-
-    def _remove(self):
-        pool = self.pool()
-        if self.uid is not None:
-            remote_id, ref_count = pool._pushed_immutable_objects.pop(self.uid)
-            if ref_count > 1:
-                pool._pushed_immutable_objects[self.remote_id] = (remote_id, ref_count - 1)
-            else:
-                mpi.call(mpi.remove_object, remote_id)
-        else:
-            mpi.call(mpi.remove_object, self.remote_id)
+    def _remove_object(self, remote_id):
+        mpi.call(mpi.remove_object, remote_id)
 
 
 def _worker_call_function(function, *args, **kwargs):
@@ -136,21 +87,6 @@ def _worker_map_function(payload, function, **kwargs):
 
     if mpi.rank0:
         return list(chain(*result))
-
-
-def _split_into_chunks(count, *args):
-    lens = map(len, args)
-    min_len = min(lens)
-    max_len = max(lens)
-    assert min_len == max_len
-    chunk_size = max_len // count + (1 if max_len % count > 0 else 0)
-
-    def split_arg(arg):
-        while arg:
-            chunk, arg = arg[:chunk_size], arg[chunk_size:]
-            yield chunk
-    chunks = tuple(list(split_arg(arg)) for arg in args)
-    return chunks
 
 
 def _setup_worker():
